+++ CHANGES.md
@@ -1,12 +1,38 @@
 # Changes
 
+## [1.0.0] - 2019-05-xx
+
 ### Add
 
+* Add `ServiceRequest::set_payload()` method.
+
+* Add `test::TestRequest::set_json()` convenience method to automatically
+  serialize data and set header in test requests.
+
+### Changes
+
+* Drop an unnecessary `Option<_>` indirection around `ServerBuilder` from `HttpServer`. #863
+
+### Fixed
+
+* Clear http requests pool on app service drop #860
+
+
+## [1.0.0-rc] - 2019-05-18
+
+### Add
+
+* Add `Query<T>::from_query()` to extract parameters from a query string. #846
 * `QueryConfig`, similar to `JsonConfig` for customizing error handling of query extractors.
 
 ### Changes
 
 * `JsonConfig` is now `Send + Sync`, this implies that `error_handler` must be `Send + Sync` too.
+
+### Fixed
+
+* Codegen with parameters in the path only resolves the first registered endpoint #841
+
 
 ## [1.0.0-beta.4] - 2019-05-12
 
@@ -18,6 +44,7 @@
 
 * `App::configure` take an `FnOnce` instead of `Fn`
 * Upgrade actix-net crates
+
 
 ## [1.0.0-beta.3] - 2019-05-04